refactor: created type for user without password, moved user storage to another directory
