+++ src/users/entities/safe-user.entity.ts
@@ -1,0 +1,3 @@
+import { User } from './users.entity';
+
+export type SafeUser = Omit<User, 'password'>;
