+++ src/users/storages/users.storage.ts
@@ -1,7 +1,9 @@
 import { Injectable } from '@nestjs/common';
 import { randomUUID } from 'crypto';
-import { CreateUserDto } from './dto/create.user.dto';
-import { User } from './users.entity';
+import { omit } from 'src/utils/omit';
+import { CreateUserDto } from '../dto/create-user.dto';
+import { SafeUser } from '../entities/safe-user.entity';
+import { User } from '../entities/users.entity';
 
 @Injectable()
 export class UsersStorage {
@@ -11,28 +13,32 @@ export class UsersStorage {
     this.users = new Map();
   }
 
-  public create(user: CreateUserDto): User {
+  public create(user: CreateUserDto): SafeUser {
     const id = randomUUID();
     const createdAt = Date.now();
     const updatedAt = createdAt;
     const version = 1;
     const newUser = { ...user, id, createdAt, updatedAt, version };
     this.users.set(id, newUser);
-    const userWithoutPassword = { ...newUser };
-    delete userWithoutPassword.password;
 
-    return userWithoutPassword;
+    return this.getById(id);
   }
 
-  public getAll(): User[] {
-    return Array.from(this.users).map(([id, user]) => user);
+  public getAll(): SafeUser[] {
+    return Array.from(this.users.values()).map((user) =>
+      omit(user, 'password'),
+    );
   }
 
-  public getById(id: string): User {
+  public getById(id: string): SafeUser {
+    return omit(this.users.get(id), 'password');
+  }
+
+  public getByIdWithPassword(id: string): User {
     return this.users.get(id);
   }
 
-  public update(id: string, user: Partial<User>): User {
+  public update(id: string, user: Partial<User>): SafeUser {
     const userForUpdate = this.users.get(id);
 
     if (!userForUpdate) {
@@ -44,10 +50,7 @@ export class UsersStorage {
 
     this.users.set(id, { ...userForUpdate, ...user, updatedAt, version });
 
-    const userWithoutPassword = this.users.get(id);
-    delete userWithoutPassword.password;
-
-    return userWithoutPassword;
+    return this.getById(id);
   }
 
   public delete(id: string): boolean {
