+++ src/users/users.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
-import { UsersStorage } from './users.storage';
+import { UsersStorage } from './storages/users.storage';
 
 @Module({
   controllers: [UsersController],
