+++ src/users/users.service.ts
@@ -1,27 +1,27 @@
 import { ForbiddenException, Injectable } from '@nestjs/common';
-import { CreateUserDto } from './dto/create.user.dto';
-import { UpdatePasswordDto } from './dto/update.password.dto';
-import { User } from './users.entity';
-import { UsersStorage } from './users.storage';
+import { CreateUserDto } from './dto/create-user.dto';
+import { UpdatePasswordDto } from './dto/update-password.dto';
+import { SafeUser } from './entities/safe-user.entity';
+import { UsersStorage } from './storages/users.storage';
 
 @Injectable()
 export class UsersService {
   constructor(private readonly collection: UsersStorage) {}
 
-  create(createUserDto: CreateUserDto): User {
+  create(createUserDto: CreateUserDto): SafeUser {
     return this.collection.create(createUserDto);
   }
 
-  getAll() {
+  getAll(): SafeUser[] {
     return this.collection.getAll();
   }
 
-  getById(id: string): User {
+  getById(id: string): SafeUser {
     return this.collection.getById(id);
   }
 
-  update(id: string, data: UpdatePasswordDto): User {
-    const user = this.collection.getById(id);
+  update(id: string, data: UpdatePasswordDto): SafeUser {
+    const user = this.collection.getByIdWithPassword(id);
 
     if (!user) {
       return;
